Tidy up of the code and other items.
