+++ docs/code_ref/index.rst
@@ -8,6 +8,8 @@ Code Reference
    :maxdepth: 2
 
    sunkit_image
-   radial
-   utils
    enhance
+   noise
+   radial
+   trace
+   utils
