+++ docs/code_ref/trace.rst
@@ -1,0 +1,7 @@
+*****
+trace
+*****
+
+This package implements tracing routines for solar physics data.
+
+.. automodapi:: sunkit_image.trace
