+++ sunkit_image/noise.py
