+++ sunkit_image/tests/test_noise.py
