--- sunkit_image/utils/__init__.py
@@ -1,2 +1,0 @@
-from .noise import *
-from .utils import *
